Initial commit: Basic list
This is the first step toward creating a simple testing framework in
Rust. It is insipired by Kevlin Henney's LHR and Catch for C++. I spent
some time understanding how to create a singly linked list in Rust
because I'm also learning this language.

The test must be organised in the specification-proposal fashion, as a
starting point.