Structure the tests with macros
The basic structure of the test should be dictated by the proposal
format: test_case is the root which defines the set up for all the
proposals and use proposal! macros to define the test which now can have
their name and a block that defines their behaviour.

Use a require macro instead of assertion to let the program execute
subsequent tests after a failure.