Hierarchical propositions in the test framework
With exploiting the nested blocks and scopes in Rust it is possible to
create 'setUp' code block in the parent test case. Which means that
there is no need to create classes or anything like that, and the test
flows as a specification properly.

There are still issues with the hierarchical reporting as the
result of a 'branch' proposal will be reported at the end of all leaf
proposal, which is quiet out of context there.

The system works by running two phases, a dry run once, when it collects
all the proposals. At the moment this data is not used properly.