Sorting out the test traversal
Found the big issue with the 'requirement' macro which I couldn't nest
previously. It turns out that it was the reuse of the template parameter
names caused some issue.

Change the macro parameter names to be more explicit of the role they
play. Then arrange the requirement macro under the 'section' macro so it
can use the variables defined in those blocks. This means that the
'proposals' can be removed from the system as the sections nesting just
fine and the condition to be 'leaf' is to have 'require' macros.

Print the relevant information to the screen more akin to catch++.
changed.